Added operators and fixed warnings

+++ STRUCTURES/src/Hash/HashTable.cpp
@@ -3,8 +3,8 @@
 
 namespace STRUCTS {
 
-    HashTable::HashTable(int capacity) : capacity(capacity), size(0), table(std::make_unique<std::unique_ptr<HashNode>[]>(capacity)) {
-        for (int i = 0; i < capacity; ++i) {
+    HashTable::HashTable(int capacity) : capacity(capacity), size(0), table(std::make_unique<std::unique_ptr<HashNode>[]>(static_cast<size_t>(capacity))) {
+        for (size_t i = 0; i < static_cast<size_t>(capacity); ++i) {
             table[i] = nullptr;
         }
     }
@@ -18,20 +18,20 @@ namespace STRUCTS {
     }
 
     void HashTable::resize(int new_capacity) {
-        std::unique_ptr<std::unique_ptr<HashNode>[]> new_table = std::make_unique<std::unique_ptr<HashNode>[]>(new_capacity);
-        for (int i = 0; i < new_capacity; ++i) {
+        std::unique_ptr<std::unique_ptr<HashNode>[]> new_table = std::make_unique<std::unique_ptr<HashNode>[]>(static_cast<size_t>(new_capacity));
+        for (size_t i = 0; i < static_cast<size_t>(new_capacity); ++i) {
             new_table[i] = nullptr;
         }
 
-        for (int i = 0; i < capacity; ++i) {
+        for (size_t i = 0; i < static_cast<size_t>(capacity); ++i) {
             std::unique_ptr<HashNode>& node = table[i];
             while (node != nullptr) {
                 int new_index = node->key % new_capacity;
                 std::unique_ptr<HashNode> next_node = std::move(node->next);
 
                 // Move the node to the new table
-                node->next = std::move(new_table[new_index]);
-                new_table[new_index] = std::move(node);
+                node->next = std::move(new_table[static_cast<size_t>(new_index)]);
+                new_table[static_cast<size_t>(new_index)] = std::move(node);
                 node = std::move(next_node);
             }
         }
@@ -41,12 +41,12 @@ namespace STRUCTS {
     }
 
     void HashTable::insert(int key, int value) {
-        if (size >= capacity * 0.75) {
+        if (size >= static_cast<int>(capacity * 0.75)) {
             resize(capacity * 2);
         }
 
         int index = hashFunction(key);
-        std::unique_ptr<HashNode>& node = table[index];
+        std::unique_ptr<HashNode>& node = table[static_cast<size_t>(index)];
         while (node != nullptr) {
             if (node->key == key) {
                 node->value = value;
@@ -56,14 +56,14 @@ namespace STRUCTS {
         }
 
         auto new_node = std::make_unique<HashNode>(key, value);
-        new_node->next = std::move(table[index]);
-        table[index] = std::move(new_node);
+        new_node->next = std::move(table[static_cast<size_t>(index)]);
+        table[static_cast<size_t>(index)] = std::move(new_node);
         ++size;
     }
 
     void HashTable::remove(int key) {
         int index = hashFunction(key);
-        std::unique_ptr<HashNode>& node = table[index];
+        std::unique_ptr<HashNode>& node = table[static_cast<size_t>(index)];
         std::unique_ptr<HashNode>* prev = &node;
 
         while (node != nullptr && node->key != key) {
@@ -81,7 +81,7 @@ namespace STRUCTS {
 
     int HashTable::search(int key) const {
         int index = hashFunction(key);
-        const std::unique_ptr<HashNode>& node = table[index];
+        const std::unique_ptr<HashNode>& node = table[static_cast<size_t>(index)];
 
         for (const HashNode* n = node.get(); n != nullptr; n = n->next.get()) {
             if (n->key == key) {
@@ -94,7 +94,7 @@ namespace STRUCTS {
 
     bool HashTable::contains(int key) const {
         int index = hashFunction(key);
-        const std::unique_ptr<HashNode>& node = table[index];
+        const std::unique_ptr<HashNode>& node = table[static_cast<size_t>(index)];
 
         for (const HashNode* n = node.get(); n != nullptr; n = n->next.get()) {
             if (n->key == key) {
@@ -107,7 +107,7 @@ namespace STRUCTS {
 
     int& HashTable::operator[](int key) {
         int index = hashFunction(key);
-        std::unique_ptr<HashNode>& node = table[index];
+        std::unique_ptr<HashNode>& node = table[static_cast<size_t>(index)];
         while (node != nullptr) {
             if (node->key == key) {
                 return node->value;
@@ -117,14 +117,14 @@ namespace STRUCTS {
 
         // If key not found, insert a new node with default value 0
         insert(key, 0);
-        return table[hashFunction(key)]->value;
+        return table[static_cast<size_t>(hashFunction(key))]->value;
     }
 
     bool HashTable::operator==(const HashTable& other) const {
         if (capacity != other.capacity || size != other.size) {
             return false;
         }
-        for (int i = 0; i < capacity; ++i) {
+        for (size_t i = 0; i < static_cast<size_t>(capacity); ++i) {
             const std::unique_ptr<HashNode>& node1 = table[i];
             const std::unique_ptr<HashNode>& node2 = other.table[i];
             const HashNode* n1 = node1.get();
